Rename sentry package to reporting
